Fixed like button functionality
set_like could not find the correct like. set_like is now executed before set_spot. @like is the argument of the path that is specified in the link (i.e., the link from the upvote or downvote button). Also, the @like was not saved when loading a spots#show page before so that the likes#set_spot could not find any existing like in the database for the user-spot combination.